Internal: Adjust functionality of upload SVG control [ED-12192] (#23937)
